requirements.txt
Django==1.10
django-bcrypt==0.9.2
django-bootstrap3==12.0.3
django-jquery==3.1.0
py-bcrypt==0.4
